Gallons Per Fish Checker
Simple project which detects how many fish would be suitable for an aquarium. This takes user input for gallons, current fish, and fish they plan to add to the aquarium. Using the total number of fish and the number of gallons the tank hold,  it returns whether the gallon amount is good enough for the number of fish.